Changed the file structure for ec-CC computations a bit.
